oci-runtime-audit-remediation-v2: full audit remediation
Critical bug fixes:
- Fix PodmanImageParser.parse_digest_from_pull (noise-prefix skip tuple)
- Fix PodmanImageParser.parse_inspect null RepoTags
- Add Podman auth-error patterns
- Remove fd < 1000 guard from ProcessPipeReader
- Add ValueError catch to CliTransport stdin thread
- Fix CliStreamingTransport cancellation ordering
- Guard logs(follow) GeneratorExit in finally

High-severity:
- Timeout-aware poll loop in all 3 transports
- Fix RuntimeFactoryConfig type annotations
- Unify timeout type to float | None
- Fix RuntimePreference docstring

Build error contract:
- build() wraps ParsingError -> ImageRuntimeError(from e)

stream_output:
- Route through OutputStream callbacks when stream_output=True

Composition root consolidation:
- Move manager construction from provider to factory
- Remove RuntimeProvider.create_managers

Deep immutability:
- Add _freeze_mapping/_freeze_sequence helpers
- dict->MappingProxyType, list->tuple for domain types

Test fidelity:
- Add timeout/cancellation/PTY/tar/logs tests
- Add manager-level ImagePullAccessDeniedError tests
- Remove misleading concurrency tests
- Fix build-command key ordering
- Remove/replace non-meaningful tests
- Replace sys.stdout.buffer patching with OutputStream injection
- Replace time.sleep cancellation tests with deterministic

Documentation:
- Update ARCHITECTURE.md for all changes
- Fix output_stream.py docstring
- Add remediation log entry

840 tests pass.

+++ src/shared/oci-runtime/src/oci_runtime/ports/transport.py
@@ -10,7 +10,7 @@ class Transport(ABC):
         self,
         command: list[str],
         *,
-        timeout: int | None = None,
+        timeout: float | None = None,
         input_data: bytes | None = None,
         cancel_token: CancellationToken | None = None,
     ) -> RawExecResult: ...
